pyproject.toml
[project]
name = "acid-chess"
dynamic = ["version"]
authors = [
  { name="Bernhard Janetzki", email="boerni@gmail.com" },
]
description = "The Chess Computer for nerds, by nerds."
readme = "README.md"
requires-python = ">=3.9"
keywords = ["python", "chess", "ai", "neural-network", "image-processing", "pytorch"]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "albumentations==1.3.0",
    "CairoSVG==2.6.0",
    "chess==1.9.4",
    "gTTS==2.3.1",
    "pyobjc==9.0.1; sys_platform=='darwin'",
    "imutils==0.5.4",
    "labelme2coco==0.2.2",
    "lightning==2.0.0",
    "numpy==1.24.2",
    "Pillow==9.4.0",
    "playsound==1.3.0",
    "pycocotools==2.0.6",
    "PySide6==6.4.2",
    "python_dateutil==2.8.2",
    "pytorch_lightning==1.9.4",
    "qimage2ndarray==1.10.0",
    "torchmetrics==0.11.4",
    "torchvision==0.15.1",
]

[project.urls]
"Homepage" = "https://github.com/ierror/acid-chess"
"Bug Tracker" = "https://github.com/ierror/acid-chess/issues"

[project.scripts]
acid = "acid.app.__main__:main"

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["acid"]

[tool.setuptools.dynamic]
version = {attr = "acid.__version__"}

[tool.black]
line-length = 120

[tool.isort]
line_length = 120
profile = "black"

[tool.flake8]
ignore = ['E501', 'W503']